+++ examples/hello_world_embassy/src/main.rs
@@ -191,11 +191,12 @@ async fn main(spawner: embassy_executor::Spawner) {
 
     let app = make_static!(make_app());
 
-    let config = make_static!(picoserve::Config {
-        start_read_request_timeout: Some(Duration::from_secs(5)),
-        read_request_timeout: Some(Duration::from_secs(1)),
-        write_timeout: Some(Duration::from_secs(1)),
-    });
+    let config = make_static!(picoserve::Config::new(picoserve::Timeouts {
+        start_read_request: Some(Duration::from_secs(5)),
+        read_request: Some(Duration::from_secs(1)),
+        write: Some(Duration::from_secs(1)),
+    })
+    .keep_connection_alive());
 
     let shared_control = SharedControl(make_static!(Mutex::new(control)));